Compiling

+++ Drivers/ESP01/espUsartEncoding.h
@@ -45,7 +45,7 @@ typedef struct
 ESPPKT_DECODEEnum ESPPKT_Decode(ESPQ_Typedef *queue, ESPPKT_RxPacket_TD *packet);
 ESP_Result ESPPKT_Encode(uint8_t *data, uint16_t length, ESPQ_Typedef *queue);
 ESP_Result ESPPKT_EncodeStart(ESPQ_Typedef *queue, uint16_t totallength, uint8_t *partData, uint16_t partlength, uint32_t *crc);
-ESP_Result ESPPKT_EncodePart(ESPQ_Typedef *queue, uint8_t *partData, uint16_t partlength, uint16_t *crc);
-ESP_Result ESPPKT_EncodeEnd(ESPQ_Typedef *queue, uint8_t *data, uint16_t length, uint16_t *crc);
+ESP_Result ESPPKT_EncodePart(ESPQ_Typedef *queue, uint8_t *partData, uint16_t partlength, uint32_t *crc);
+ESP_Result ESPPKT_EncodeEnd(ESPQ_Typedef *queue, uint8_t *data, uint16_t length, uint32_t *crc);
 
 #endif /* BEN_PACKET_H */
